requirements.txt
spotipy~=2.23.0
Flask-Session~=0.4.0
docopt~=0.6.2
jupyterlab~=3.4.4
seaborn~=0.11.2
dash~=2.6.1
pandas~=1.3.5
dash-bootstrap-components~=1.2.1
